+++ blockstack-sdk/src/androidTest/java/org/blockstack/android/sdk/BlockstackSessionAuthProfileTest.kt
@@ -5,6 +5,9 @@ import android.preference.PreferenceManager
 import android.support.test.rule.ActivityTestRule
 import android.support.test.runner.AndroidJUnit4
 import kotlinx.coroutines.experimental.runBlocking
+import org.blockstack.android.sdk.model.Profile
+import org.blockstack.android.sdk.model.Proof
+import org.blockstack.android.sdk.model.toBlockstackConfig
 import org.blockstack.android.sdk.test.TestActivity
 import org.hamcrest.CoreMatchers.`is`
 import org.hamcrest.CoreMatchers.nullValue