+++ blockstack-sdk/src/main/java/org/blockstack/android/sdk/model/SessionData.kt
@@ -1,4 +1,4 @@
-package org.blockstack.android.sdk
+package org.blockstack.android.sdk.model
 
 import org.json.JSONObject