+++ example/src/main/java/org/blockstack/android/MainActivity.kt
@@ -16,9 +16,15 @@ import kotlinx.android.synthetic.main.content_main.*
 import kotlinx.coroutines.experimental.android.UI
 import kotlinx.coroutines.experimental.async
 import org.blockstack.android.sdk.*
+import org.blockstack.android.sdk.model.GetFileOptions
+import org.blockstack.android.sdk.model.PutFileOptions
+import org.blockstack.android.sdk.model.UserData
+import org.blockstack.android.sdk.model.toBlockstackConfig
 import org.jetbrains.anko.coroutines.experimental.bg
 import java.io.ByteArrayOutputStream
 import java.net.URL
+
+private const val username = "dev_android_sdk.id.blockstack";
 
 @SuppressLint("SetTextI18n")
 class MainActivity : AppCompatActivity() {
@@ -149,7 +155,7 @@ class MainActivity : AppCompatActivity() {
         }
 
         getStringFileFromUserButton.setOnClickListener { _ ->
-            val username = "dev_android_sdk.id.blockstack";
+
             val zoneFileLookupUrl = URL("https://core.blockstack.org/v1/names")
             fileFromUserContentsTextView.text = "Downloading file from other user..."
             blockstackSession().lookupProfile(username, zoneFileLookupURL = zoneFileLookupUrl) { profileResult ->
@@ -200,7 +206,6 @@ class MainActivity : AppCompatActivity() {
 
         getUserAppFileUrlButton.setOnClickListener { _ ->
             getUserAppFileUrlText.text = "Getting url ..."
-            val username = "dev_android_sdk.id.blockstack";
             val zoneFileLookupUrl = "https://core.blockstack.org/v1/names"
             blockstackSession().getUserAppFileUrl(textFileName, username, "https://flamboyant-darwin-d11c17.netlify.com", zoneFileLookupUrl) {
                 runOnUiThread {
@@ -227,6 +232,20 @@ class MainActivity : AppCompatActivity() {
             }, { countResult ->
                 Log.d(TAG, "files count " + if (countResult.hasValue) {countResult.value} else {countResult.error})
             })
+        }
+
+        getNameInfoButton.setOnClickListener { _ ->
+            getNameInfoText.text = "Getting info ..."
+            blockstackSession().network.getNameInfo(username) {
+                runOnUiThread {
+                    Log.d(TAG, it.value?.json.toString())
+                    getNameInfoText.text = if (it.hasValue) {
+                        it.value?.json.toString()
+                    } else {
+                        it.error
+                    }
+                }
+            }
         }
 
         if (intent?.action == Intent.ACTION_VIEW) {